chore: interface 외부 사용 타입 분리
